Automatically expand dialog
Proposed change to dialog resize behavior (for discussion!)
* Make dialog non-resizeable
* Make a VBox rather than scroll pane the main container
* Update dialog height dynamically when options are expanded

This appears to behave quite well for me, although there is a little bit of jumpiness in the vertical position during resize animations. Main question is whether this is better or worse than enabling the user to resize (it may not be so nice for people with smaller screens).