Update message label on object selection
Not the prettiest code... but this tries to maintain counts of selected detections and annotations, and use these (alongside other bindings) to show sensible messages to the user.

If things are in an invalid state, then the run button will be disabled and a warning should be shown (by default in red). Otherwise, the run button will be enabled and a count of the relevant selected objects shown.

This also solves the issue of empty space appearing in the dialog if there is no warning to display.